prepering user info
