+++ app/src/main/java/com/example/fitlane/FirstFragment.kt
@@ -11,6 +11,10 @@ import com.example.fitlane.databinding.FragmentFirstBinding
 import android.app.Service
 //import android.content.Intent
 import android.os.IBinder
+import android.util.Log
+import android.widget.Button
+import android.widget.Toast
+import com.google.firebase.auth.FirebaseAuth
 import java.util.*
 /**
  * A simple [Fragment] subclass as the default destination in the navigation.
@@ -36,6 +40,21 @@ class FirstFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?)
     {
         super.onViewCreated(view, savedInstanceState)
+
+        binding.TimerStart.setOnClickListener{
+            Toast.makeText( activity, "hej", Toast.LENGTH_SHORT).show()
+            val userS:String = FirebaseAuth.getInstance().currentUser?.uid.toString()
+            val emailS:String = FirebaseAuth.getInstance().currentUser?.email.toString()
+            var textS:String = "info:\n"
+            textS+= userS+"\n"
+            textS+= emailS+"\n"
+            Toast.makeText( getActivity(), textS, Toast.LENGTH_SHORT).show()
+
+            Log.d("text", "aaaa")
+            var t: Button = view.findViewById(R.id.Timer_start)
+            t.text= ("aaaa")
+
+        }
     }
